models updated after meeting
