+++ app/models/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+from pdps import PDP
+from statuses import Status
+from tasks import Task
+from type import Type
+from users import User
