+++ app/models/base.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, Date, DateTime, Integer
+from sqlalchemy import Column, Date, Integer
 from sqlalchemy.orm import declarative_base, declared_attr
 
 
@@ -12,7 +12,6 @@ class BaseModel:
         return cls.__name__.lower()
 
     id = Column(Integer, primary_key=True)
-    created = Column(DateTime, default=datetime.now)
 
 
 Base = declarative_base(cls=BaseModel)
@@ -21,5 +20,5 @@ Base = declarative_base(cls=BaseModel)
 class AbstractDatesModel(Base):
     """Абстрактная модель с датами начала и окончания"""
     __abstract__ = True
-    starting_date = Column(Date(), nullable=False)
+    starting_date = Column(Date(), default=datetime.date.now)
     deadline = Column(Date(), nullable=False)
